requirement.txt
fastapi
uvicorn
pdfplumber
sentence-transformers
faiss-cpu
aiohttp
numpy
requests